requirements.txt
absl-py==0.8.0
astor==0.8.0
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
cloudpickle==1.2.2
cycler==0.10.0
decorator==4.4.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.3.0
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==2.2.4
mock==3.0.5
networkx==2.2
numpy==1.16.5
pathlib==1.0.1
Pillow==6.1.0
protobuf==3.9.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
scikit-image==0.14.5
scipy==1.2.2
six==1.12.0
subprocess32==3.5.4
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2